+++ packages/providers/notion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/notion
 
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies [a8ec08c]
+  - @llamaindex/core@0.6.16
+
 ## 0.1.14
 
 ### Patch Changes