+++ packages/providers/storage/mongodb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/mongodb
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [a8ec08c]
+  - @llamaindex/core@0.6.16
+
 ## 0.0.30
 
 ### Patch Changes